moved create_tables in a separate file (db_init.py), removed database.py
